created working Makefile which would also trigger libft make, added test file and linked libraries properly

+++ libftprintf/runtests.sh
@@ -1,2 +1,2 @@
-cc -Wextra -Wall -Werror testprintf/printf_test.c && ./a.out
-rm ./a.out
+cc -Wextra -Wall -Werror test_file.c -L. -lftprintf -L./libft -lft -I./libft && ./a.out
+[ -e ./a.out ] && rm ./a.out
